Move the existing product code to the product directory
Only change was a typo in "unknown".
