+++ lms/models/__init__.py
@@ -25,7 +25,6 @@ from lms.models.lti_registration import LTIRegistration
 from lms.models.lti_role import LTIRole
 from lms.models.lti_user import LTIUser, display_name
 from lms.models.oauth2_token import OAuth2Token
-from lms.models.product import Product
 from lms.models.rsa_key import RSAKey
 from lms.models.user import User
