+++ tests/unit/conftest.py
@@ -7,7 +7,8 @@ from pyramid import testing
 from pyramid.request import apply_request_extensions
 
 from lms.db import SESSION
-from lms.models import ApplicationSettings, Product
+from lms.models import ApplicationSettings
+from lms.product import Product
 from tests import factories
 from tests.conftest import TEST_SETTINGS, get_test_database_url
 from tests.unit.services import *  # pylint: disable=wildcard-import,unused-wildcard-import
